Improve User add_game functionality
Add add_game_by_title instead of handling String and Game cases
in add_game. add_game_by_title checks for an existing game with the
same title, and if so calls add_game with the existing game. If not,
it creates a new game with that title, but formats the title nicely.
This should prevent a bunch of crappy differences in titleization.