Add a simple (not nice looking) implementation of remove buttons
I want to come up with an interface that works well on the desktop
and on mobile. The idea with this current interface was to hide the
remove buttons initially and then show them with a menu click, then
hide them again once the user clicked one of them and removed the
game they wanted to remove. Right now they're just shown all the time
and it looks really bad on mobile (the buttons go below the titles).
Look into using an icon font (X) icon instead of the text Remove and
make it not break to the line below on mobile.