User can click Edit List to delete games
Replace delete buttons with (X) icons, hidden initially
Add 'Edit List' link to top nav bar. When the user clicks this
it shows the (X) icons by each game. Clicking those will delete
the associated game and refresh the list. When in edit mode the
'Edit List' link changes to 'Done Editing'. When the user clicks
that the (X) icons are hidden again.